Make or break
Looks like repo hosted an older version of the plugin (dated 2013, so around the date of presenting the poster).

What I'm uploading is the newer version (dated 2015) that we're hosting at https://github.com/AMUFacultyOfEnglish/MoodlePlugin. I think this is the version that we used on our Moodle (code mentions fixes for Moodle 2.9)

So, I'm either uploading the latest version, or I'm breaking it. Which one is it? You'll never know.